remove shorts

+++ prime-dotnet-webapi/Models/BusinessEventType.cs
@@ -9,10 +9,10 @@ namespace Prime.Models
     [Table("BusinessEventTypeLookup")]
     public sealed class BusinessEventType : BaseAuditable, ILookup<int>
     {
-        public const short STATUS_CHANGE_CODE = 1;
-        public const short EMAIL_CODE = 2;
-        public const short NOTE_CODE = 3;
-        public const short ADMIN_CLAIM_CODE = 4;
+        public const int STATUS_CHANGE_CODE = 1;
+        public const int EMAIL_CODE = 2;
+        public const int NOTE_CODE = 3;
+        public const int ADMIN_CLAIM_CODE = 4;
 
         [Key]
         public int Code { get; set; }
